+++ server/deps/repository_adapters/occurrence_repository_adapter.py
@@ -3,6 +3,7 @@ from typing import Annotated
 from fastapi import Depends
 
 from server.deps.authenticate import UserDep
+from server.deps.conflict_checker import ConflictCheckerDep
 from server.deps.owned_building_ids import OwnedBuildingIdsDep
 from server.deps.repository_adapters.classroom_repository_adapter import (
     ClassroomRepositoryDep,
@@ -14,15 +15,18 @@ from server.deps.session_dep import SessionDep
 from server.models.database.occurrence_db_model import Occurrence
 from server.models.database.schedule_db_model import Schedule
 from server.models.http.requests.allocate_request_models import AllocateSchedule
+from server.repositories.intentional_conflict_repository import (
+    IntentionalConflictRepository,
+)
 from server.repositories.occurrence_repository import OccurrenceRepository
 from server.services.security.classrooms_permission_checker import (
-    classroom_permission_checker,
+    ClassroomPermissionChecker,
 )
 from server.services.security.occurrence_permission_checker import (
-    occurrence_permission_checker,
+    OccurrencePermissionChecker,
 )
 from server.services.security.schedule_permission_checker import (
-    schedule_permission_checker,
+    SchedulePermissionChecker,
 )
 
 
@@ -34,6 +38,7 @@ class OccurrenceRepositoryAdapter:
         user: UserDep,
         classroom_repo: ClassroomRepositoryDep,
         schedule_repo: ScheduleRepositoryDep,
+        conflict_checker: ConflictCheckerDep,
     ):
         self.owned_building_ids = owned_building_ids
         self.session = session
@@ -41,6 +46,12 @@ class OccurrenceRepositoryAdapter:
         self.owned_building_ids = owned_building_ids
         self.classroom_repo = classroom_repo
         self.schedule_repo = schedule_repo
+        self.checker = ClassroomPermissionChecker(user=user, session=session)
+        self.occurrence_checker = OccurrencePermissionChecker(
+            user=user, session=session
+        )
+        self.schedule_checker = SchedulePermissionChecker(user=user, session=session)
+        self.conflict_checker = conflict_checker
 
     def get_all(self) -> list[Occurrence]:
         occurrences = OccurrenceRepository.get_all_on_buildings(
@@ -50,15 +61,21 @@ class OccurrenceRepositoryAdapter:
 
     def get_by_id(self, id: int) -> Occurrence:
         occurrence = OccurrenceRepository.get_by_id(id=id, session=self.session)
-        occurrence_permission_checker(self.user, occurrence, self.session)
+        self.occurrence_checker.check_permission(occurrence)
         return occurrence
 
-    def allocate_schedule(self, schedule_id: int, classroom_id: int) -> Schedule:
+    def allocate_schedule(
+        self, schedule_id: int, classroom_id: int, intentional_conflict: bool = False
+    ) -> Schedule:
         schedule = self.schedule_repo.get_by_id(schedule_id)
         classroom = self.classroom_repo.get_by_id(classroom_id)
         OccurrenceRepository.allocate_schedule(
-            user=self.user, schedule=schedule, classroom=classroom, session=self.session
+            user=self.user,
+            schedule=schedule,
+            classroom=classroom,
+            session=self.session,
         )
+
         self.session.commit()
         self.session.refresh(schedule)
         return schedule
@@ -69,7 +86,7 @@ class OccurrenceRepositoryAdapter:
     ) -> None:
         for pair in schedule_classroom_pairs:
             schedule = self.schedule_repo.get_by_id(pair.schedule_id)
-            schedule_permission_checker(self.user, schedule, self.session)
+            self.schedule_checker.check_permission(schedule)
             if pair.classroom_id == -1:
                 OccurrenceRepository.remove_schedule_allocation(
                     user=self.user, schedule=schedule, session=self.session
@@ -77,14 +94,28 @@ class OccurrenceRepositoryAdapter:
                 continue
 
             classroom = self.classroom_repo.get_by_id(pair.classroom_id)
-            classroom_permission_checker(self.user, classroom, self.session)
+            self.checker.check_permission(classroom)
 
-            OccurrenceRepository.allocate_schedule(
+            occurrences = OccurrenceRepository.allocate_schedule(
                 user=self.user,
                 schedule=schedule,
                 classroom=classroom,
                 session=self.session,
             )
+            if pair.intentional_conflict and pair.intentional_occurrence_ids:
+                intentional_occurrences = OccurrenceRepository.get_by_ids(
+                    ids=pair.intentional_occurrence_ids, session=self.session
+                )
+                occurrences.sort(key=lambda x: x.date)
+                for occurrence in occurrences:
+                    for intentional_occurrence in intentional_occurrences:
+                        if occurrence.conflicts_with(intentional_occurrence):
+                            IntentionalConflictRepository.create(
+                                first_occurrence=occurrence,
+                                second_occurrence=intentional_occurrence,
+                                session=self.session,
+                            )
+
         self.session.commit()
 
     def remove_schedule_allocation(self, schedule_id: int) -> Schedule:
